separates type_of_force table (known as Tags) into its own folder
