+++ api/tags/tags.test.js
@@ -1,0 +1,3 @@
+/* eslint-disable */
+const db = require('../../data/db-config');
+const Tags = require('./tagsModel');
